+++ game/tiles.py
@@ -1,4 +1,5 @@
 import random
+from diccionario import DATA
 
 class TooMuchTiles (Exception):
     pass
@@ -6,7 +7,7 @@ class TooMuchTiles (Exception):
 class TooMuchTilesPut(Exception):
     pass
 
-TOTALTILES=98
+TOTALTILES=100
 
 class Tile:
     def __init__(self, letter, value):
@@ -15,36 +16,10 @@ class Tile:
 
 class TileBag:
     def __init__(self):
-        self.tiles =  [
-            Tile('A', 1), Tile('A', 1), Tile('A', 1), Tile('A', 1), Tile('A', 1), Tile('A', 1), Tile('A', 1), Tile('A', 1), Tile('A', 1), Tile('A', 1), Tile('A', 1), Tile('A', 1),
-            Tile('E', 1), Tile('E', 1), Tile('E', 1), Tile('E', 1), Tile('E', 1), Tile('E', 1), Tile('E', 1), Tile('E', 1), Tile('E', 1), Tile('E', 1), Tile('E', 1), Tile('E', 1),
-            Tile('I', 1), Tile('I', 1), Tile('I', 1), Tile('I', 1), Tile('I', 1), Tile('I', 1),
-            Tile('L', 1), Tile('L', 1), Tile('L', 1), Tile('L', 1),
-            Tile('N', 1), Tile('N', 1), Tile('N', 1), Tile('N', 1), Tile('N', 1),
-            Tile('O', 1), Tile('O', 1), Tile('O', 1), Tile('O', 1), Tile('O', 1), Tile('O', 1), Tile('O', 1), Tile('O', 1), Tile('O', 1),
-            Tile('R', 1), Tile('R', 1), Tile('R', 1), Tile('R', 1), Tile('R', 1),
-            Tile('S', 1), Tile('S', 1), Tile('S', 1), Tile('S', 1), Tile('S', 1), Tile('S', 1),
-            Tile('T', 1), Tile('T', 1), Tile('T', 1), Tile('T', 1),
-            Tile('U', 1), Tile('U', 1), Tile('U', 1), Tile('U', 1), Tile('U', 1),
-            Tile('D', 2), Tile('D', 2), Tile('D', 2), Tile('D', 2), Tile('D', 2),
-            Tile('G', 2), Tile('G', 2), 
-            Tile('B', 3), Tile('B', 3),
-            Tile('C', 3), Tile('C', 3), Tile('C', 3), Tile('C', 3),
-            Tile('M', 3), Tile('M', 3),
-            Tile('P', 3), Tile('P', 3),
-            Tile('F', 4),
-            Tile('H', 4), Tile('H', 4),
-            Tile('V', 4),
-            Tile('Y', 4),
-            Tile('CH', 5),
-            Tile('Q', 5),
-            Tile('J', 8),
-            Tile('LL', 8),
-            Tile('Ñ', 8),
-            Tile('RR', 8),
-            Tile('X', 8),
-            Tile('Z', 10)
-        ]
+        self.tiles=[]
+        for i in DATA:
+            for j in range(i.get('quantity')):
+                self.tiles.append(Tile(i.get("letter"),i.get("value")))
         random.shuffle(self.tiles)
  
     def draw_tiles(self,cantidad):
@@ -72,3 +47,15 @@ class TileBag:
     def tiles_remaining(self):
         return len(self.tiles)
 
+class JokerTile(Tile):
+    def __init__(self, letter, value):
+        super().__init__(letter, value)
+
+
+    def chooseLetter(self):
+        tilebag=TileBag()
+        letter_joker=str(input('Elija una letra para el Joker: '))
+        if letter_joker in DATA.get('letter'):
+            self.tiles.replace('?',letter_joker)
+            
+