+++ CHANGELOG.md
@@ -26,3 +26,27 @@ Development.
 ### Issues
 
 - Anything at the moment
+
+## Status
+
+Development.
+
+## (0.0.2) V 18-08-23
+
+### In Progress
+
+- Test Joker Tile.
+
+### Added
+
+- Joker Tile Class.
+- Diccionary For Tiles.
+- Choose Joker Tile Method.
+
+### Changed
+
+-Tile list change by a for method
+
+### Issues
+
+- Choose Joker Tile Test. Not working.